Updated UML class diagram. Created unit tests for Chunk class and implemented Chunk class. Refactored Table class.
