+++ health/src/test/java/com/health/TableTest.java
@@ -90,55 +90,4 @@ public class TableTest {
 		
 		assertNull(actual);
 	}
-
-	@Test(expected=NullPointerException.class)
-	public void addRecord_givenRecordNull_throwsNullPointerException() {
-		Table table = arrangeTable();
-		
-		table.addRecord(null);
-	}
-	
-	@Test
-	public void addRecord_givenRecord_addsRecord() {
-		Table table = arrangeTable();
-		Record record = arrangeRecord();
-				
-		table.addRecord(record);
-
-		assertThat(table.getRecords(), hasItem(record));
-	}
-	
-	@Test
-	public void removeRecord_givenRecordExists_removesRecord() {
-		Table table = arrangeTable();
-		Record record = arrangeRecord();		
-		table.addRecord(record);
-		
-		table.removeRecord(record);
-
-		assertThat(table.getRecords(), not(hasItem(record)));
-	}
-	
-	@Test
-	public void removeRecord_givenRecordDoesNotExist_doesNotRemoveRecords() {
-		Table table = arrangeTable();
-		Record record1 = arrangeRecord();	
-		Record record2 = arrangeRecord();		
-		table.addRecord(record1);
-		
-		table.removeRecord(record2);
-
-		assertThat(table.getRecords(), hasItems(record1));
-	}
-
-	@Test
-	public void removeRecord_givenRecordNull_doesNotRemoveRecords() {
-		Table table = arrangeTable();
-		Record record = arrangeRecord();		
-		table.addRecord(record);
-		
-		table.removeRecord(null);
-
-		assertThat(table.getRecords(), hasItems(record));
-	}
 }
